# specs/001-crud-empleados/contracts/empleados.openapi.yaml
openapi: 3.0.3
info:
    title: API CRUD Empleados
    version: 1.0.0
    description: Contrato del CRUD de empleados con autenticación HTTP Basic.
servers:
    - url: http://localhost:8080
security:
    - basicAuth: []
paths:
    /api/departamentos:
        post:
            summary: Crear departamento
            operationId: crearDepartamento
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/DepartamentoCreateRequest"
            responses:
                "201":
                    description: Departamento creado
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/DepartamentoResponse"
                "400":
                    description: Datos inválidos
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "409":
                    description: Clave duplicada
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
        get:
            summary: Listar departamentos
            operationId: listarDepartamentos
            responses:
                "200":
                    description: Lista de departamentos
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: "#/components/schemas/DepartamentoResponse"

    /api/departamentos/{clave}:
        get:
            summary: Obtener departamento por clave
            operationId: obtenerDepartamentoPorClave
            parameters:
                - in: path
                  name: clave
                  required: true
                  schema:
                      type: string
            responses:
                "200":
                    description: Departamento encontrado
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/DepartamentoResponse"
                "404":
                    description: Departamento no encontrado
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
        put:
            summary: Actualizar departamento
            operationId: actualizarDepartamento
            parameters:
                - in: path
                  name: clave
                  required: true
                  schema:
                      type: string
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/DepartamentoUpdateRequest"
            responses:
                "200":
                    description: Departamento actualizado
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/DepartamentoResponse"
                "400":
                    description: Datos inválidos
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "404":
                    description: Departamento no encontrado
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
        delete:
            summary: Eliminar departamento
            operationId: eliminarDepartamento
            parameters:
                - in: path
                  name: clave
                  required: true
                  schema:
                      type: string
            responses:
                "204":
                    description: Operación idempotente de eliminación (exista o no el recurso)

    /api/empleados:
        post:
            summary: Crear empleado
            operationId: crearEmpleado
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/EmpleadoCreateRequest"
            responses:
                "201":
                    description: Empleado creado
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/EmpleadoResponse"
                "400":
                    description: Datos inválidos
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "409":
                    description: Clave duplicada
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
        get:
            summary: Listar empleados
            operationId: listarEmpleados
            responses:
                "200":
                    description: Lista de empleados
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: "#/components/schemas/EmpleadoResponse"

    /api/empleados/{clave}:
        get:
            summary: Obtener empleado por clave
            operationId: obtenerEmpleadoPorClave
            parameters:
                - in: path
                  name: clave
                  required: true
                  schema:
                      type: string
            responses:
                "200":
                    description: Empleado encontrado
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/EmpleadoResponse"
                "404":
                    description: Empleado no encontrado
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
        put:
            summary: Actualizar empleado
            operationId: actualizarEmpleado
            parameters:
                - in: path
                  name: clave
                  required: true
                  schema:
                      type: string
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/EmpleadoUpdateRequest"
            responses:
                "200":
                    description: Empleado actualizado
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/EmpleadoResponse"
                "400":
                    description: Datos inválidos
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "404":
                    description: Empleado no encontrado
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "409":
                    description: Conflicto por concurrencia o versión
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
        delete:
            summary: Eliminar empleado
            operationId: eliminarEmpleado
            parameters:
                - in: path
                  name: clave
                  required: true
                  schema:
                      type: string
            responses:
                "204":
                    description: Operación idempotente de eliminación (exista o no el recurso)

components:
    securitySchemes:
        basicAuth:
            type: http
            scheme: basic
    schemas:
        EmpleadoCreateRequest:
            type: object
            required: [clave, nombre, direccion, telefono]
            properties:
                clave:
                    type: string
                    minLength: 1
                    maxLength: 100
                nombre:
                    type: string
                    minLength: 1
                    maxLength: 100
                direccion:
                    type: string
                    minLength: 1
                    maxLength: 100
                telefono:
                    type: string
                    minLength: 1
                    maxLength: 100
                departamentoClave:
                    type: string
                    minLength: 1
                    maxLength: 100
        EmpleadoUpdateRequest:
            type: object
            required: [nombre, direccion, telefono, version]
            properties:
                nombre:
                    type: string
                    minLength: 1
                    maxLength: 100
                direccion:
                    type: string
                    minLength: 1
                    maxLength: 100
                telefono:
                    type: string
                    minLength: 1
                    maxLength: 100
                version:
                    type: integer
                    format: int64
                    minimum: 0
                departamentoClave:
                    type: string
                    minLength: 1
                    maxLength: 100
        EmpleadoResponse:
            type: object
            required: [clave, nombre, direccion, telefono, version]
            properties:
                clave:
                    type: string
                nombre:
                    type: string
                    description: Valor persistido en longitud variable (sin relleno artificial), máximo 100.
                direccion:
                    type: string
                    description: Valor persistido en longitud variable (sin relleno artificial), máximo 100.
                telefono:
                    type: string
                    description: Valor persistido en longitud variable (sin relleno artificial), máximo 100.
                version:
                    type: integer
                    format: int64
                departamentoClave:
                    type: string
                    description: Clave del departamento asociado si existe
        ErrorResponse:
            type: object
            required: [code, message]
            properties:
                code:
                    type: string
                message:
                    type: string
        DepartamentoCreateRequest:
            type: object
            required: [clave, nombre]
            properties:
                clave:
                    type: string
                    minLength: 1
                    maxLength: 100
                nombre:
                    type: string
                    minLength: 1
                    maxLength: 100
        DepartamentoUpdateRequest:
            type: object
            required: [nombre]
            properties:
                nombre:
                    type: string
                    minLength: 1
                    maxLength: 100
        DepartamentoResponse:
            type: object
            required: [clave, nombre]
            properties:
                clave:
                    type: string
                nombre:
                    type: string