# src/main/java/com/dsw01/practica02/web/dto/DepartamentoCreateRequest.java
package com.dsw01.practica02.web.dto;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;

public class DepartamentoCreateRequest {

    @NotBlank
    @Size(max = 100)
    private String clave;

    @NotBlank
    @Size(max = 100)
    private String nombre;

    public String getClave() {
        return clave;
    }

    public void setClave(String clave) {
        this.clave = clave;
    }

    public String getNombre() {
        return nombre;
    }

    public void setNombre(String nombre) {
        this.nombre = nombre;
    }
}