# src/main/java/com/dsw01/practica02/web/dto/DepartamentoResponse.java
package com.dsw01.practica02.web.dto;

public class DepartamentoResponse {

    private String clave;
    private String nombre;

    public String getClave() {
        return clave;
    }

    public void setClave(String clave) {
        this.clave = clave;
    }

    public String getNombre() {
        return nombre;
    }

    public void setNombre(String nombre) {
        this.nombre = nombre;
    }
}