# src/main/java/com/dsw01/practica02/web/dto/ErrorResponse.java
package com.dsw01.practica02.web.dto;

public class ErrorResponse {

    private String code;
    private String message;

    public ErrorResponse(String code, String message) {
        this.code = code;
        this.message = message;
    }

    public String getCode() {
        return code;
    }

    public String getMessage() {
        return message;
    }
}